Migrate JDBC sampler onto the SamplerExtensionPoint mechanism
JDBCSamplerProvider's inline Sampler lambda moves verbatim into a new
JDBCSampler.query(String) method; the provider becomes a thin wrapper
that binds via SamplerExtensionPoint.bind(). JDBCSamplerOptions.query
now carries a full expression (query("SELECT 1")) instead of bare SQL,
so JDBCSamplerProviderIT, MySQLSamplerProviderIT, and
PostgreSQLSamplerProviderIT are updated to match; JDBCSamplerPlugin is
untouched since it only ever forwards the string through.
